backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pytesseract
pillow
openpyxl
pandas
python-dotenv
google-generativeai
